REFACTOR: move userManagement to services directory
