+++ src/controllers/UserController.java
@@ -2,17 +2,17 @@ package src.controllers;
 
 import javax.swing.JOptionPane;
 
-import src.models.UserManagementModel;
+import src.services.UserManagementService;
 import src.models.UserModel;
 import src.utils.UserSearchType;
 import src.views.UserViews;
 
 public class UserController {
-	UserManagementModel userManagementModel = UserManagementModel.getInstance();
+	UserManagementService userManagementService = UserManagementService.getInstance();
 	UserViews userView = new UserViews();
 
 	private void handleFind(String queryId, UserSearchType userSearchType) {
-		UserModel foundUser = userManagementModel.findUser(queryId, userSearchType);
+		UserModel foundUser = userManagementService.findUser(queryId, userSearchType);
 		if (foundUser == null) {
 			JOptionPane.showMessageDialog(null, "Usuario no encontrado.");
 			return;
@@ -58,7 +58,7 @@ public class UserController {
 				name = JOptionPane.showInputDialog(null, "Ingresa el Nombre", "");
 				id = JOptionPane.showInputDialog(null, "Ingresa el ID", "");
 
-				userManagementModel.addUser(new UserModel(name, id));
+				userManagementService.addUser(new UserModel(name, id));
 				parsedSelectedOption = "Atrás";
 				break;
 
@@ -66,7 +66,7 @@ public class UserController {
 				id = JOptionPane.showInputDialog(null, "Ingresa el ID", "");
 				newName = JOptionPane.showInputDialog(null, "Ingresa el Nuevo Nombre", "");
 				newID = JOptionPane.showInputDialog(null, "Ingresa el Nuevo ID", "");
-				userManagementModel.updateUser(newName, id, newID);
+				userManagementService.updateUser(newName, id, newID);
 				parsedSelectedOption = "Atrás";
 				break;
 			case "Eliminar un Usuario (ID)":
